+++ src/main/java/com/sanxia/salesManagement/system/model/OffInfo.java
@@ -1,5 +1,6 @@
 package com.sanxia.salesManagement.system.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class OffInfo {
@@ -13,7 +14,7 @@ public class OffInfo {
 
     private Date endTime;
 
-    private Long totalDays;
+    private BigDecimal totalDays;
 
     private String reason;
 
@@ -57,11 +58,11 @@ public class OffInfo {
         this.endTime = endTime;
     }
 
-    public Long getTotalDays() {
+    public BigDecimal getTotalDays() {
         return totalDays;
     }
 
-    public void setTotalDays(Long totalDays) {
+    public void setTotalDays(BigDecimal totalDays) {
         this.totalDays = totalDays;
     }
